ci: don't build the merge ref
This has proven to be buggy, and misleading. When force-pushing a PR
branch, the merge head doesn't get updated reliably, and the CI run thus
builds the wrong thing, old code.

Additionally, the BK UI will claim it was building the correct commit,
which confuses even more.

This adds an info message with the commit being built, for extra
visibility.

Signed-off-by: Stephan Renatus <srenatus@chef.io>