ci: ensure changes to master don't get lost
The previous commit had a potential issue: since this "git rebase"
action happens on every build (from the pre-command BK hook), it checks
origin/master many times, and what that resolves to could change in the
time between jobs -- especially for something like the long-running
'build' job, and the ones of step 2.

Now, before, this would just be swallowed, and could lead to confusion.
Now, we'll end up with multiple annotations. It's still be confusing,
but not invisible.

Signed-off-by: Stephan Renatus <srenatus@chef.io>