ci: don't build the merge ref, try to rebase ourselves (#583)
* ci: don't build the merge ref

This has proven to be buggy, and misleading. When force-pushing a PR
branch, the merge head doesn't get updated reliably, and the CI run thus
builds the wrong thing, old code.

Additionally, the BK UI will claim it was building the correct commit,
which confuses even more.

This adds an info message with the commit being built, for extra
visibility.

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* ci: rebase onto master

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* ci: tweak build annotations

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* ci: ensure changes to master don't get lost

The previous commit had a potential issue: since this "git rebase"
action happens on every build (from the pre-command BK hook), it checks
origin/master many times, and what that resolves to could change in the
time between jobs -- especially for something like the long-running
'build' job, and the ones of step 2.

Now, before, this would just be swallowed, and could lead to confusion.
Now, we'll end up with multiple annotations. It's still be confusing,
but not invisible.

Signed-off-by: Stephan Renatus <srenatus@chef.io>