session-service: allow for multiple relay states
Before, the following could happen:

- tab A: get logged out, sent to dex -- relay state 1
- tab B: get logged out, sent to dex -- relay state 2, overwrites relay
  state 1
- login in tab A, be sent back to session-service
- relay state 2 != 1, error out

Now, we don't bind a single relay state to the session (which would be
overwritten when a second logout occours), but multiple relay states.
To achieve this, we don't store a string key in the session (with key
"relay_state"), but multiple booleans, with key
"relay_state_<RELAY_STATE>". Since we only need to check whether a state
we get back from dex matches a known relay state, and never need to
retrieve all relay states, that works.

This is a shortcut to not thinking too hard about whether we _need_ the
relay state checking.

Signed-off-by: Stephan Renatus <srenatus@chef.io>