session-service/server: index client_state by relay_state
Before, login attempts in two different tabs would have succeeded, but
they would both, when redirect to automate-ui, have ended up on the same
UI page.

To avoid that, we use a variant of the relay_state trick to store
multiple client_state values, i.e., the UI route the client is going to
be redirected to.

Signed-off-by: Stephan Renatus <srenatus@chef.io>