automate-gateway: add ?pretty parameter to every grpc-gateway method
Now, if you query, for example, either one of

    https://a2-dev.test/apis/iam/v2beta/policies?pretty
    https://a2-dev.test/apis/iam/v2beta/policies?pretty=1
    https://a2-dev.test/apis/iam/v2beta/policies?pretty=true
    https://a2-dev.test/apis/iam/v2beta/policies?pretty=please

you'll get a pretty-printed JSON response ONLY for methods provided
through grpc-gateway.

Also note that the same can be achieved by supplying a request header:

    curl -H "Accept: application/json+pretty" ...

since this is what this change does under the hood. If such a header
has been provided in the API request, AND that value is NOT */*, the

    Query "?pretty" => Header "Accept: application/json+pretty"

translation will happen for grpc-gateway-provided routes.

Signed-off-by: Stephan Renatus <srenatus@chef.io>