Add top-level products.meta file
The products.meta file describes which products we can deploy, along
with which packages belong to those products. This PR also adds a
package.meta that can be added for any package in its component
directory. With this file, we can describe things about the file like if
it's a data service, what binlinks it should install, etc.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>