automate-gateway: add ?pretty parameter to every grpc-gateway method (#614)
* automate-gateway: add ?pretty parameter to every grpc-gateway method

Now, if you query, for example, either one of

    https://a2-dev.test/apis/iam/v2beta/policies?pretty
    https://a2-dev.test/apis/iam/v2beta/policies?pretty=1
    https://a2-dev.test/apis/iam/v2beta/policies?pretty=true
    https://a2-dev.test/apis/iam/v2beta/policies?pretty=please

you'll get a pretty-printed JSON response ONLY for methods provided
through grpc-gateway.

Also note that the same can be achieved by supplying a request header:

    curl -H "Accept: application/json+pretty" ...

since this is what this change does under the hood. If such a header
has been provided in the API request, AND that value is NOT */*, the

    Query "?pretty" => Header "Accept: application/json+pretty"

translation will happen for grpc-gateway-provided routes.

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* automate-chef-io: use ?pretty in simple case for auth[nz] docs

Simple cases are those where we don't project, like, 'jq .id', or select
a subset.

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* a2-api-integration: add ?pretty assertion

Yes, the assertion is hideous. Good there's only of these.

Signed-off-by: Stephan Renatus <srenatus@chef.io>