compliance export: set start time to beg of day of end time
on the backend side of things, we are ignoring the start time for the
export endpoint, and simulating the behaviour here, where we say
start time is beg of day of end time. we would like to change this to
respect the provided start time for api users, so we need to adjust the
ui to send start time as beg of day of end time (instead of beg of day
from 10 days ago) so that we can maintain the behaviour of the ui export.

Signed-off-by: Victoria Jeffrey <vjeffrey@chef.io>