[SHAME] Pass signleton PolicyRefresher up from policy server instance to project server instance and use it call Refresh to get the cache correct before kicking off the reindex
This is obviously gross but to initialize the signleton, we need the store instance which is also created in the NewPolicyServer call. Should we be sharing a single v2 store instance between all v2 GRPC servers? Is this good enough until we tackle the big refactor of this part of the code scheduled to be done soon? Open to whatever suggestions here.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>