Adding InSpec version to Compliance page search bar (#660)
Adding the InSpec version to the Compliance page search bar. During testing this addition we noticed that for suggestions '.' characters are treated differently. The problem was that the Elasticsearch tokenizer was not including the '.' in the tokens. To fix the problem we have to add the "punctuation" term to the "autocomplete_tokenizer".

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>