[event-feed-service] Apply golangci-lint
This project was already very close to being lint clean, so I've
applied the linter to it and added a studio function for running the
linter.

This unfortunately duplicates some of the functionality of we have in
Makefile.common_go, but I didn't want to wade into the task of fixing
the curious blend of Makefiles, shell scripts, and studio functions
that various parts of the repository uses.

We may want to have go_component_static_tests run the linter; however,
I'll defer that until all the projects that use
go_component_static_tests are passing the linter.

Signed-off-by: Steven Danna <steve@chef.io>