dev/acceptance envs: remove saml lies and deal with the fallout (#726)
* terraform/test-envs: adjust descriptions re: SAML or no SAML

They call claim to do SAML, when really only six of them have a record
in our Okta instance.

* terraform/test-envs: add SAML variable, set it where okta records exist
* terraform/test-envs: depend on SAML var for dex config

Learnt about local variables here:

https://medium.com/@pavloosadchyi/terraform-patterns-and-tricks-i-use-every-day-117861531173

No idea how else to do that; a ternary operator with so many lines in it
didn't seem right, so I haven't tried that.

* terraform/test-envs: fix README slack channel
* e2e/cypress: make sense of local-user-only login form; skip "SSO" specs for these
* terraform/test-envs: update README
* terraform/test-envs: tag machines with SAML set up

This is step 1 for using the data to drive the e2e runs: we'll need to
know if "SSO" is a thing for the specs we run.

Since it seems that data.json's tags only contains the values, we're using

    X-SAML = "saml"

and not

    X-SAML = "true"

* terraform/test-envs: add tag for machines we run e2e/cypress against
* e2e/cypress: drive script from tags

The idea is that whichever machine is tagged "e2e" will get cypress run
against it. If the machine also has a "saml" tag, the SSO tests are
*not* skipped.

* e2e/cypress: make script work with jq 1.5
* e2e/cypress: fix if-then-else

Still hate bash.