Bug fix: DeleteRule was actually broken
Previouly after staging deletion of an applied rule,
GetRule would return the applied rule
rather than the staged deleted rule.
It returned the applied rule because `query_staged_or_applied_rule`
saw no results from staging due to absence of conditions.
(Note that even changing the join to an outer join to force a result
did not work--then the row.Scan failed because of a nil value
for the condition Values array.)

"Returning the staged deleted rule" means it comes back
with "not found for reasons of marked for deletion".
The problem was that we violated the built-in assumption that rules
will always have conditions. Always. Even for a staged deletion.
So the fix is to give that staged deleted rule a dummy condition.
Not pretty, but the alternative is to change the fundamental assumption
and allow empty condition lists.

Signed-off-by: michael sorens <msorens@chef.io>