authz: fix flaky delete test
the problem was that we were trying to copy the applied rule db_id
into the staged rule table and getting a conflict, because there already
existed a staged rule with that db_id.

now, we don't copy the applied rule db_id over at all. the newly
staged-for-deletion rule gets the next staged db_id in sequence,
just like when we create or update a rule.

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>