authz: new approach to condition validation
before, we included the rule type on the condition object
so we could validate that condition attributes matched the correct rule type.
(right now, that just means Event rules can only have Chef Server or Chef Org attributes)

since we weren't even storing the condition type in the database and it always
had to match the rule type,
it seemed unnecessary for the condition object to have a type at all

so i removed the field and moved the condition attribute validation into the
rule validation.

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>