[A2-292] Property-based tests for rules (#712)
* Revise test framework to mock ingest connections

Signed-off-by: michael sorens <msorens@chef.io>

* Starting to add new tests

Signed-off-by: michael sorens <msorens@chef.io>

* Add test for deleting rule

Signed-off-by: michael sorens <msorens@chef.io>

* Add a helper to reduce duplication

Signed-off-by: michael sorens <msorens@chef.io>

* More tests

Signed-off-by: michael sorens <msorens@chef.io>

* Refactor to allow multiple tests

Signed-off-by: michael sorens <msorens@chef.io>

* Separate test functions

Signed-off-by: michael sorens <msorens@chef.io>

* Fine-tune delete combinations

Signed-off-by: michael sorens <msorens@chef.io>

* Apply some spit and polish

Signed-off-by: michael sorens <msorens@chef.io>

* Improve test names

Signed-off-by: michael sorens <msorens@chef.io>

* One more update test

Signed-off-by: michael sorens <msorens@chef.io>

* resync after rebase with master

* Generate multiple rules instead of a single rule

Signed-off-by: michael sorens <msorens@chef.io>

* Implement using multiple rules with ListRules

Note that I needed to introduce a limit to the condtion count on rules
because ListRules was hitting the GRPC limit of 4194304

Signed-off-by: michael sorens <msorens@chef.io>

* Another day, another test... and some more cleanup

Signed-off-by: michael sorens <msorens@chef.io>

* Applying review feedback

Signed-off-by: michael sorens <msorens@chef.io>

* A bit more cleanup

Signed-off-by: michael sorens <msorens@chef.io>

* Don't neglect the return value of CreateRule

Signed-off-by: michael sorens <msorens@chef.io>

* Bug fix:Set missing rule status

Prior to this, both CreateRule and UpdateRule
returned an empty status value.

Signed-off-by: michael sorens <msorens@chef.io>

* Bug fix: DeleteRule was actually broken

Previouly after staging deletion of an applied rule,
GetRule would return the applied rule
rather than the staged deleted rule.
It returned the applied rule because `query_staged_or_applied_rule`
saw no results from staging due to absence of conditions.
(Note that even changing the join to an outer join to force a result
did not work--then the row.Scan failed because of a nil value
for the condition Values array.)

"Returning the staged deleted rule" means it comes back
with "not found for reasons of marked for deletion".
The problem was that we violated the built-in assumption that rules
will always have conditions. Always. Even for a staged deletion.
So the fix is to give that staged deleted rule a dummy condition.
Not pretty, but the alternative is to change the fundamental assumption
and allow empty condition lists.

Signed-off-by: michael sorens <msorens@chef.io>

* Another pass at giving better names of properties.

Signed-off-by: michael sorens <msorens@chef.io>