Adding a context deadline and lower case sort value (#750)
Preventing the Client Runs page's node export from going into an infinite loop. To cause the infinite loop when using the node export, a user needs to have nodes with the same name that has uppercase characters. This bug is caused by using the Elasticsearch SearchAfter API with non-matching curser values. When requesting the second page of values the last value from the first page's node name is used in the request. For the node-state index, the node name values are matched as lower-case, but the code was sending the values with uppercase. When this happens, SearchAfter API will return the first page again.

I have not seen that this infinite loop timeout in any way. Therefore a context deadline of one minute was added to prevent this from happening again.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>