Update missing db tables on service updates
We were lacking some data update on service health events,
currently we only update its health, update strategy and
package identifier, but it is missing to update the supervisor,
deployment and service-group data,specifically the application
name, environment, site and fqdn.

This commit is updating the ingestion to update all missing tables on
any service update.

Signed-off-by: Salim Afiune <afiune@chef.io>