Remove unused @angular/http package
We had converted from @angular/http to @angular/common/http
some time ago but apparently the actual package had never been removed.
The presence of the package is an impediment for
the impending Angular 8 upgrade.
`ng update` complained about an incompatible peer dependency for it,
which I suppose is true, but it was not really pointing me towards
the correct issue.

Signed-off-by: michael sorens <msorens@chef.io>