[pg-gateway] Set server-fin timeout
When a go process dies because of a signal or a panic, defer'd
database closes may not execute. In this case, HAProxy will keep the
client side connection in CLOSE-WAIT and the server-side connection in
ESTAB. Setting `server-fin` times out the server-side connection
quickly, avoiding a leaked process. Without this, the process would
not close for 5 minutes until the tunnel timeout is hit.

Signed-off-by: Steven Danna <steve@chef.io>