Add update strategy to the channel badge (#839)
The update strategy tells users how the services will be updated among a group. Habitat supports three update strategies: `none`, `rolling`, and `at-once`. 

Showing the data along with the channel name helps users understand where the service will get an update from, and how it will update.

This change is including the update strategy for a service instance in the channel badge, in the format of `[channel name]: [update strategy name]` (there is a space after the `:`)

When there is NO channel and NO update strategy, display `NO CHANNEL: NONE`

Signed-off-by: Salim Afiune <afiune@chef.io>