[pg-gateway] Set server-fin timeout (#847)
When a go process dies because of a signal or a panic, defer'd
database closes may not execute. In this case, HAProxy will keep the
client side connection in CLOSE-WAIT and the server-side connection in
ESTAB. Setting `server-fin` times out the server-side connection
quickly, avoiding a leaked process. Without this, the process would
not close for 5 minutes until the tunnel timeout is hit.

I can't think of a compelling reason to keep these connections around
any longer than we need to since the postgresql shutdown procedure is:

> The normal, graceful termination procedure is that the frontend
  sends a Terminate message and immediately closes the connection. On
  receipt of this message, the backend closes the connection and
  terminates.

So the server is expecting an immediate close from the client once it
has sent a Terminate and the server doesn't typically send a
response.

Signed-off-by: Steven Danna <steve@chef.io>