Remove transactions from task workers (#739)
Most of the design reviewers were concerned with the prospect of
long-live transactions. This PR removes long-lived transactions and
replaces it with a periodic `update` from the worker thread along
with a periodic garbage collection of expired tasks.

Now, when a task worker is lost, the workflow will receive a workflow
event indicated that the task was lost whereas previously the task
would potentially get picked up by another worker.

This is moving us from an "at least once" model to an "at most once"
model for a single enqueued task.  If a workflow wants to retry the 
lost task, it can re-enqueue the task.

We've also started moving our SQL functions from plain SQL to
plpgsql.

Signed-off-by: Steven Danna <steve@chef.io>
Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>
Co-authored-by: Jay Mundrawala <jmundrawala@chef.io>