[ci] Add a few more log lines
Based on other work in #90, I think that the stop and start
chef-automate is likely causing a long delay in the tests because at
we effectively have to wait for a timeout to happen before that will
actually complete.

It also looks like verify-packages might occasionally be very slow.
This adds some logging so that we can verify that in the CI logs.

Signed-off-by: Steven Danna <steve@chef.io>