[ci] Restructure tests to work around bug (#873)
There is currently a bug related to a bad interaction between the
periodic cert check and our shutdown logic. The code is very
entangled and it is going to take some time to sort out safely. For
now, I think modifying the cert generation test may lead us to time
out less frequently.

Signed-off-by: Steven Danna <steve@chef.io>