[deployment-service] Avoid hang on shutdown
target.Stop() is a asynchronous command because eventually `hab sup
term` will terminate the process itself.  Here, we were passing a
context into Stop that would be cancelled at the end of the function.
In most cases, this meant that `hab sup term` never got a chance to
request that the supervisor terminated and thus the service would not
be shut down until we hit the systemd timeout.

Signed-off-by: Steven Danna <steve@chef.io>