Bring up services in the order they are required during mitigation
If deployment service takes down something core like postgres, it first
has to take down all the services that connect to it. Those services
aren't brought up until the end of the run converge part, which was
causing automate-pg-gateway to come up really late in the game. This in
turn causes a bunch of healt checks to fail and causes general slowness
as hab has to service a bunch of services that will just keep dying.

This change will reload any stopped services due to the mitigation when
they are required.

I left the RestartServices at the end. It just call `hab svc start`. I'm
not sure it is needed as we run through the list of services in a
topologically sorted order.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>