[integration] Allow users to specify filters that are always used.
This makes a new variable:

`test_diagnostics_pre_upgrade_filters`

which should be used on any diagnostics run on an _old_ version.  The
existing variable

`test_diagstics_filters`

should be applied to *every* invocation of `diagnostics run`.

I've also removed some uses of skip-for-deep-upgrades that didn't make
sense to me. The tests should show if they are needed.

Finally, it isn't clear to me that we are properly managing the
`~remove-this-tag-after-merge` tag anymore.

Signed-off-by: Steven Danna <steve@chef.io>