a2-iam-v2-integration: update projects: "*" field of policy statements
This is returned now. The question we'll have to discuss is if this is
acceptable. I think it might be, since we've already included a projects
field in the responses for iam v2 (not v2.1).  The distinction just
isn't as clear-cut as I'd like it to be right now, but it might not be
worth fixing now we're so deep into this.

Signed-off-by: Stephan Renatus <srenatus@chef.io>