shellcheck: add "-s bash"
It seems that the shebang of .studiorc is ignored when it's sourced by
some other script, and we're asking shellcheck to follow sourced scripts
(-ax). So, instead we're explicitly declaring the shell to assume to be
bash.

It does make sense, too, as without it, shellcheck will flag some
non-Posix-sh stuff in that file.

Signed-off-by: Stephan Renatus <srenatus@chef.io>