[cereal] Periodically cleanup workflow results
Currently, the workflow results table grows without bound. This
imposes a limit of 10,000 rows that we attempt to enforce
periodically.

To avoid having to think too hard about the implications of DELETE and
our default transaction level, we take a postgresql advisory lock.

Note that we "over delete" to try to avoid deleting on every
interation once we hit the limit.

Signed-off-by: Steven Danna <steve@chef.io>