address review feedback
key changes:

- decided to back off using the request polling after apply rules,
because the GET /apply-rules status endpoint doesn't seem to report
an accurate complete time with such a small dataset.
(or so the high flakiness seemed to tell me)
i'd like to get to the bottom of this,
but i'll save that for a follow-up PR.

- refactored the generate admin API token code block into a custom
Cypress command so now it can be used in all the tests!

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>