[lib/cereal] Ignore missing task results
The risk with this error return location and other similar locations
is that by simply returning the error, we will be aborting the
transaction, effectively placing the same workflow event on the queue.

This means that if the error is _not_ a transient error, we are
completely stuck processing the same workflow event over and over
again.

Here, I've handled at least one non-transient error. Any other error
will continue to be retried. I think that some of this might be made
easier if we moved away from QueryRowContext since then Scan errors
and other errors wouldn't be intermingled.

Signed-off-by: Steven Danna <steve@chef.io>