deployment-service: reconcile this code with automate-cli usage (#951)
For reasons that might have disappeared from everyone's memory, the d-s
code was also what drove `chef-automate admin-token`. So, I've removed
any "diagnostics"-related wording from the V2 policy data.

Note that `chef-automate admin-token` first checks if the system uses
v2, so that code path we've recently added for diagnostics should NOT be
executed in that scenario. It still felt cleaner to remove the
"diagnostics" words.

Furthermore, this changes `GenerateAdminToken` so that in the
diagnostics+iam-v2 use case, it can be run twice: once before the
upgrade, once after; without triggering any policy conflicts.

Before this commit, the second run would have not failed completely, but
the already existing policy would not have referenced the proper token
ID. Now, the v2 code is analogous to the v1 code: one policy per token.
While we could do something more clever, this is enough for the limited
use case I'm after.

If we did do something more clever, we might as well make `chef-automate
admin-token` not freak out if A2 is running on IAMv2.

Signed-off-by: Stephan Renatus <srenatus@chef.io>