Update `chef-automate dev verify-packages` for use in CI (#969)
This changes the `chef-automate dev verify-packages` such that:

- Missing FILES files is now a handled error, allowing other packages
  to be checked.

- The package name is first in the output, making it a little easier
  to parse.

With these changes, I've replaced the old shell function that we have
to detect broken packages in the deep_upgrade tests and replaced it
with our verify-packages command. This will hopefully help catch a
broader set of package corruption issues during those tests. [1]

I've also refactored the VerifyError struct into multiple structs to
make it more immediately clear which kind of error you are hoping to
return.

[1] The package corruption issues in those tests are
unavoidable. Those tests exercise upgrades from old versions of
deployment-service. Older versions of deployment-service (a) used
versions of Hab that did not install packages atomically and (b)
didn't avoid package installs when a restart for upgrade was likely.

Signed-off-by: Steven Danna <steve@chef.io>