Parallelization of bulk publisher for high loads (#927)
Created a message distributor to send messages to multiple bulk publishers. The distributor will create a set number of bulk publishers that the messages will be sent to. The messages are sent to the first publisher until it is filled. Then the overflow messages are sent to the second publisher until it full and so on with the other publisher until they are all full. When they are all full the distributor will wait and try to send the messages again. This algorithm can be easily changed by updating the distributeMessage function.

The reason behind filling one of the publishers before sending messages to the other ones is to increase the size of the bulk messages send. With a round-robin or equal distribution approach, all the bulk publishers will be sending small messages with the overhead of more connections to elasticsearch. For example, we would rather send 100 messages with one elasticsearch connection than 4 connections with 25 messages apiece. Once the load on the system is high enough that the bottleneck is publishing the bulk messages. The second bulk publisher's inbox will fill up and the second bulk publisher will start sending bulk messages in parallel.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>