Handle empty statement projects for v2 and v2.1 (#886)
* Handle empty statement projects for v2 and v2.1

We were not dealing with empty statement projects correctly.
For v2, it should be coerced to a single element list containing "*"
indicating all projects.
For v2.1, it should fail as an error.

Signed-off-by: michael sorens <msorens@chef.io>

* Add unit tests for v2.1

Signed-off-by: michael sorens <msorens@chef.io>

* diagnostics/iamv2: use tag to disable for non-iam-v2 integration tests

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* inspec: move iam-v2 controls into a2-iam-v2-integration

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* integration: add IAM v2 diagnostics run

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* backup_ontop.sh: skip iam-v2, too

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* a2-iam-v2-integration: update projects: "*" field of policy statements

This is returned now. The question we'll have to discuss is if this is
acceptable. I think it might be, since we've already included a projects
field in the responses for iam v2 (not v2.1).  The distinction just
isn't as clear-cut as I'd like it to be right now, but it might not be
worth fixing now we're so deep into this.

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* integration/iam_v2_only_diagnostics: skip iam v1 tests

These would only pass if we migrated the v1 policies. We don't.

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* iam_v2_only_diagnostics: remove cruft, enable upgrades

Is it that easy? We'll see.

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* integration: add comment, rename test

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* diagnostics/iamv2: disable more iam-v1

Disabled user/team diagnostics when explicitly testing v2 -- there are
no corresponding v2 (or v2.1) tests yet!

Signed-off-by: Stephan Renatus <srenatus@chef.io>