[lib/cereal] Temporarily pause processing of wayward workflows (#947)
The processWorkflow event loop has a number of places where errors can
occur. In many cases, the most likely cause of the error is some
transient network problem. In these cases, logging the error, and
retrying may make sense.

Unfortunately, for the case of an unknown workflow event, no amount of
retrying will help. The event types are part of our source code, if we
don't know the event now, we won't know it 10 milliseconds from now
either.

This is especially problematic because it is likely that we'll
continue to select this same workflow event, stopping work on all
other workflows.

This PR attempts to address the problem by temporarily marking such
workflows as "wayward" and removing them from the list of workflows to
select. After some timeout, we start processing them again, in the
hope that the problematic event might have been processed by someone
else.

An alternative would be to drop the event completely. I think that
might make it problematic to add new workflow events in the future.
If a new workflow client is connected and uses some newly introduced
event, we likely want that client to be able to process the event.

Signed-off-by: Steven Danna <steve@chef.io>