Teach deployment-service how to backup ssl external pg (#931)
deployment-service was using platform config to get a postgres
connection. For services, this sets up everything to connect to internal
or external postgres correctly. deployment-service doesn't have all the
bits to make that work. I've added a hack to allow it to keep using the
platform config library and correctly get the postgres connection
information.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>