[automate-832] Use role db_id in iam_statements (final part) (#942)
Updated the iam_statements table to have a foreign key to role's via db_id. Also, since things are now properly relational, if you delete a role, policies will now remove the role from their statement. If removing the role would result in the statement having no role or actions, then the statement is removed. If that results in an empty custom policy, the policy is removed.

I also added a insertTestStatement helper and updated the postgres tests to use it.

Also found and fixed a few bugs:

* Resources and actions in the iam_statements didn't have NOT NULL constraints and could have resulted in null values. I've populated any NULLs with '{}' and added NOT NULL constraints.
* Fixed a bug where errors on db transactions would cause an infinite loop and crash authz-service.