cleanup HAB_CACHE_KEY_PATH cruft
All of the file copying and key importing was added before we realized
that none of that matters because the HAB_CACHE_KEY_PATH would be
shared between builders. This removes that code and instead uses
HAB_CACHE_KEY_PATH during generation to put the key in the results/
dir directly.

Signed-off-by: Steven Danna <steve@chef.io>