Add support for filtering service groups by app, env and name
To avoid a query with substring we added a new column for name suffix to service group.
Technically the name suffix is actually the name of the service group, and the alternative
would be to compose the service.service_group name upon query.

App, Env and service group name all result in complete service group returns
this allows us to continue to use the view for these filters.

Signed-off-by: kmacgugan <kmacgugan@chef.io>