token-details: Add non-blank validator
I have added this many times over the last few PRs, but I did miss a few
update forms like this.
(Both required and non-blank are needed,
even if the pattern itself disallows an empty value.)

Also, I wanted to get this much more concise pattern in the template
when working on rules but did not get to it.
So this is the first use ('nameCtrl'). Let me know what you think.

Signed-off-by: michael sorens <msorens@chef.io>