deployment-service: don't delete dependencies in conservative-gc
In conservative mode we only delete old versions of packages if they
are in our top-level packages set. However, one of our top-level
packages may also be a dependency of another top-level package.

For example, currently pg-sidecar-service depends on an older version
of automate-postgresql but we are removing the version it depends on.

Signed-off-by: Steven Danna <steve@chef.io>