Dan/fix incorrect service grouping (#1064)
A customer reported that services are not assigned to the correct groups when they have different application settings. I was able to reproduce pretty easily by running:

```
applications_publish_supervisor_message --application ONE --sup-id 1111
applications_publish_supervisor_message --application TWO --sup-id 2222
applications_publish_supervisor_message --application THREE --sup-id 3333
```

On insert of a new service, the problem was that we were only looking at service groups based on their name, whereas we needed to be checking both the name and deployment id to get the correct row back.

After applying a fix for that problem, I investigated how the system would behave if the user has existing data that is incorrect due to this insertion bug. We recently added some code to modify the application and environment names if they change, but I've discovered that it was subtly incorrect: the existing code updates the deployment and/or service group values to match what was given in the latest health check message, but this changes the app name/environment/service group name for all services associated with the deployment/service group. I've therefore changed the logic so that when a health check message indicates that a service should belong to a different service group/deployment than what is currently in the database, we create a new deployment/service group, if necessary, then move the service over to the correct deployment and service group. I have confirmed that if there is incorrect data in the database resulting from the #1000 bug, it will be repaired as new health checks come in and the update logic gets applied.

* Correctly detect when new services need a new service group/deployment
* Rework how service group/deployment changes work for existing svcs
* Remove unused deployments/service groups when moving services
* test full matrix of app/env/svc group change scenarios
* cleanup comments and debug logging
* trim the number of branches in service update
* Fix sorting test that relied on svc mis-grouping bug
* Remove needUpdate flags where unused
* cleanup comments for hab service updates
* Cleanup comments in integration test