Prepare nats client for subcribing via multiple message streams
This prepares the nats client to start receiving messages from a nats
messaging (i.e., plain NATS) subject instead of a nats-streaming
channel. In the future we may use nats-streaming again for some kinds of
messages where we want more reliable delivery; if that happens we will
need to change the message types so the receiver can sensibly handle
multiple message types, so we are deprecating the current nats-streaming
channel.

Signed-off-by: Daniel DeLeo <dan@chef.io>