pg-sidecar-service: minor error handling fixes, new build
This fixes some uses of logrus.Fatal inside a function that returns an
error to its caller.

But, this also has the side-effect of rebuilding the package. We
haven't shipped pg-sidecar-service in a while. Because of the bug
fixed in #1111 we think that many installations likely have a small
problem where not all of pg-sidecar-service's dependencies are
installed. Shipping a new version will fix since the update
dependencies will get installed on upgrade.

Signed-off-by: Steven Danna <steve@chef.io>