license-control-service: store state in postgresql (#1100)
This changes the license-control-service to use a table in PostgreSQL
rather than a file on disk to store any configured license. The goal
of this PR is to prepare the service for running multiple instances of
this service.

With this change we now keep a history of the licenses that are
applied. We are OK with the potential unbounded growth here since we
are the only people who can generate valid licenses.

Reads from this service are now much more work than they used to be as
we will re-fetch and re-parse the license from the database each time.
If this becomes an issue, we can add caching.

Signed-off-by: Steven Danna <steve@chef.io>