Add "UNRECOGNIZED" update strategy to account for some upgrade scenarios
This change accounts for the case where:
1. A new update strategy is added to the habitat supervisor
2. A user upgrades habitat and uses this update strategy
3. But does not upgrade Automate

In this case, Automate will not be able to infer the proper name of the
update strategy because the names of the update strategies are hard
coded in Automate, and even if we did not hardcode them in the Automate
codebase, we would need to infer them from the compiled protocol buffer
code, which means that in either case a new version of the applications
service needs to be built and upgraded-to before habitat messages
referencing the new update strategy can be properly understood.

Should these conditions occur, we want to label the update strategy
"UNRECOGNIZED" instead of incorrectly coercing to one of the strategies
we do know about.

Signed-off-by: Daniel DeLeo <dan@chef.io>