Remove unnecessary copies from pkg.path
These services were copying files that are only used for reading out
of package path and into the service's static path. No one I talked to
had a reason for why we were doing this.

We think maybe one of the following let us here:

1) permissions problems on files. We now test with a restrictive
default umask and it is unclear to me what permissions problem we
could have hit that the `cp` would have avoided.

2) Following a "rule of thumb" that the service should only depend on
files in `/hab/svc/SERVICE-NAME`. While I think this is a good rule
of thumb, for static data (this data could have been compiled into the
binary), I don't see the harm in reading it directly from the package
path.

Removing the copy comes with the advantage of removing another step
from service startup that could fail.

There is still calls to `cp` in our run scripts, but they are in
services that I'm not inclined to touch at the moment:

    components/automate-cs-nginx/habitat/hooks/run
    15:cp {{cfg.required_recipe.path}} {{pkg.svc_var_path}}/required_recipe

    components/automate-cs-oc-erchef/habitat/hooks/run
    56:cp {{pkg.svc_config_path}}/dark_launch_features.json {{pkg.svc_data_path}}

    components/automate-workflow-server/habitat/hooks/run
    26:  cp -a {{pkg.path}}/{{cfg.ssh_git.git_repo_template}} {{pkg.svc_files_path}}/

Signed-off-by: Steven Danna <steve@chef.io>