compliance, nodemanager: removed unused functions
While this removes code, I think it is a bit of a downside as these
functions are kind of a family of functions that make sense to travel
together. My hope in cleaning up this kind of thing between services
is so that we can start to see what kind of functionality might be
useful to share somewhere.

Signed-off-by: Steven Danna <steve@chef.io>