compliance-service: prealloc linter
This fixes all the prealloc violations and simplifies our LINTERARGS
based on changes to the main linter configuration.

The only remaining differences in compliance-service from the standard
configuration is the nakedret violations. Fixing those looks _mostly_
straightforward but there are a few with some subtle behavior so it is
probably best done as its own PR.

Signed-off-by: Steven Danna <steve@chef.io>